# perf(ext/headers): cache iterableHeaders for immutable Headers